auto obs + tags+ multi-mine-1-line
Also a bunch of refactor,
Also
