+++ config_EXAMPLE.toml
@@ -8,8 +8,9 @@ screenshot_destination = "~/Videos/OBS/SS/"
 [anki]
 sentence_audio_field = "SentenceAudio"
 picture_field = "Picture"
-source_field = "Source"
 current_game = "Japanese Game"
+custom_tags = ['JapaneseGameMiner'] # leave Empty if you dont want to add tags
+add_game_tag = true
 
 # Feature Flags
 [features]
@@ -32,3 +33,11 @@ extension = "webp" # Codec of screenshot, Recommend Keeping this as webp (Defaul
 
 [audio]
 extension = "opus" # Desired Extension/codec of Trimmed Audio, (Default opus)
+
+[obs]
+enabled = true
+start_buffer = true
+full_auto_mode = false # Automatically Create Cards when you Create in Yomi. REQUIRED for multi-card-per-voiceline
+host = "localhost"
+port = 4455
+password = "your_password_here"
