+++ util.py
@@ -1,0 +1,32 @@
+import os
+import random
+import string
+import threading
+from datetime import datetime
+
+use_previous_audio = False
+
+lock = threading.Lock()
+
+def make_unique_file_name(path):
+    split = path.rsplit('.', 1)
+    filename = split[0]
+    extension = split[1]
+    return filename + "_" + get_random_digit_string() + "." + extension
+
+
+def get_random_digit_string():
+    return ''.join(random.choice(string.digits) for i in range(9))
+
+
+def timedelta_to_ffmpeg_friendly_format(td_obj):
+    total_seconds = td_obj.total_seconds()
+    hours, remainder = divmod(total_seconds, 3600)
+    minutes, seconds = divmod(remainder, 60)
+    return "{:02}:{:02}:{:06.3f}".format(int(hours), int(minutes), seconds)
+
+
+def get_file_modification_time(file_path):
+    mod_time_epoch = os.path.getmtime(file_path)
+    mod_time = datetime.fromtimestamp(mod_time_epoch)
+    return mod_time
